refactor: split context.rs into focused submodules
Extract the monolithic 1795-line context.rs into a context/ module with
5 focused submodules, each grouping related AppContext methods:

- mod.rs (543 lines): struct definition, constructor, config/SDK methods
- wallet_lifecycle.rs (572 lines): SPV/wallet management and reconciliation
- identity_db.rs (205 lines): identity and DPNS database operations
- contract_token_db.rs (198 lines): contract and token database operations
- settings_db.rs (84 lines): settings cache and persistence
- transaction_processing.rs (231 lines): transaction finality handling

Pure extraction with no logic changes.
